+++ efootprint/api_utils/json_to_system.py
@@ -69,7 +69,7 @@ def json_to_system(system_dict, launch_system_computations=True):
             new_obj.trigger_modeling_updates = False
             for attr_key, attr_value in system_dict[class_key][class_instance_key].items():
                 if type(attr_value) == dict:
-                    new_obj.__setattr__(attr_key, json_to_explainable_object(attr_value))
+                    new_obj.__setattr__(attr_key, json_to_explainable_object(attr_value), check_input_validity=False)
                 else:
                     new_obj.__dict__[attr_key] = attr_value
 
@@ -82,15 +82,16 @@ def json_to_system(system_dict, launch_system_computations=True):
         for mod_obj_key, mod_obj in class_obj_dict[class_key].items():
             for attr_key, attr_value in list(mod_obj.__dict__.items()):
                 if type(attr_value) == str and attr_key != "id" and attr_value in flat_obj_dict.keys():
-                    mod_obj.__setattr__(attr_key, ContextualModelingObjectAttribute(flat_obj_dict[attr_value]))
+                    mod_obj.__setattr__(attr_key, ContextualModelingObjectAttribute(flat_obj_dict[attr_value]),
+                                        check_input_validity=False)
                 elif type(attr_value) == list and attr_key != "contextual_modeling_obj_containers":
                     output_val = []
                     for elt in attr_value:
                         if type(elt) == str and elt in flat_obj_dict.keys():
                             output_val.append(flat_obj_dict[elt])
-                    mod_obj.__setattr__(attr_key, ListLinkedToModelingObj(output_val))
+                    mod_obj.__setattr__(attr_key, ListLinkedToModelingObj(output_val), check_input_validity=False)
             for calculated_attribute in mod_obj.calculated_attributes:
-                mod_obj.__setattr__(calculated_attribute, EmptyExplainableObject())
+                mod_obj.__setattr__(calculated_attribute, EmptyExplainableObject(), check_input_validity=False)
 
     for obj_type in class_obj_dict.keys():
         if obj_type != "System":