+++ efootprint/builders/services/generative_ai_ecologits.py
@@ -42,7 +42,7 @@ class GenAIModel(Service):
 
         return {"model_name": {"depends_on": "provider", "conditional_list_values": values}}
 
-    def __setattr__(self, name, input_value):
+    def __setattr__(self, name, input_value, check_input_validity=True):
         if name == "provider" and self.trigger_modeling_updates:
             raise PermissionError("The provider of a GenAIModel cannot be changed after initialization. "
                                   "Use a ModelingUpdate to update both provider and a compatible model name.")