+++ efootprint/core/hardware/server.py
@@ -24,7 +24,6 @@ class Server(ServerBase):
             "server_utilization_rate": SourceValue(0.9 * u.dimensionless, Sources.HYPOTHESIS),
             "base_ram_consumption": SourceValue(0 * u.GB, Sources.HYPOTHESIS),
             "base_compute_consumption": SourceValue(0 * u.cpu_core, Sources.HYPOTHESIS),
-            "storage": Storage.ssd("Default SSD storage"),
             "fixed_nb_of_instances": EmptyExplainableObject()
         }