+++ tests/builders/hardware/test_boavizta_server_from_config.py
@@ -34,15 +34,6 @@ class TestBoaviztaServerFromConfig(unittest.TestCase):
             fixed_nb_of_instances=EmptyExplainableObject()
         )
 
-    def test_calculated_attributes_have_been_computed(self):
-        """
-        Test that the calculated attributes have been computed.
-        """
-        self.assertTrue(self.test_server.carbon_footprint_fabrication.value > 0 * u.kg)
-        self.assertTrue(self.test_server.power.value > 0 * u.W)
-        self.assertTrue(self.test_server.ram.value > 0 * u.GB)
-        self.assertTrue(self.test_server.compute.value > 0 * u.cpu_core)
-
     def test_update_cpu_config(self):
         """
         Verify that update_cpu_config() uses self.nb_of_cpu_units & self.nb_of_cores_per_cpu_unit